requirements.txt
bcrypt==3.1.3
beautifulsoup4==4.6.0
bleach==2.1.1
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.3.1
Flask-WTF==0.14.2
gunicorn==19.7.1
html-sanitizer==1.3.0
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
lxml==4.1.0
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
pew==1.1.1
Pillow==4.3.0
pipenv==8.3.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
requests==2.18.4
rsa==3.4.2
selenium==3.7.0
six==1.11.0
SQLAlchemy==1.1.14
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
webencodings==0.5.1
Werkzeug==0.12.2
WTForms==2.1
